fix migration issue

+++ config/initializers/fake_rate.rb
@@ -1,2 +1,4 @@
-ExchangeRate.destroy_all
-ExchangeRate.create({rate: 5.45})
+if(defined? ExchangeRate)
+  ExchangeRate.destroy_all
+  ExchangeRate.create({rate: 5.45})
+end
